feat: add functionality to fetch the title of a webpage when creating a URL
The changes in this commit add the ability to fetch the title of a webpage when creating a URL. This is done by adding a new function `fetchTitle` that makes an HTTP GET request to the provided URL and parses the HTML response to extract the title. The fetched title is then stored in the `Title` field of the `DocShort` struct. This allows users to have a better understanding of the content of the webpage associated with a shortened URL.